+++ java/spark-1.6.0/rdd.scala
@@ -4,10 +4,8 @@ import org.apache.spark._
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql._
 
-import scala.reflect.ClassTag
-
-class WorkerRDD[T: ClassTag](
-  parent: RDD[T],
+class WorkerRDD(
+  prev: RDD[Row],
   closure: Array[Byte],
   columns: Array[String],
   config: String,
@@ -15,22 +13,21 @@ class WorkerRDD[T: ClassTag](
   groupBy: Array[String],
   closureRLang: Array[Byte],
   bundlePath: String
-  ) extends RDD[T](parent) {
+  ) extends RDD[Row](prev) {
 
   private[this] var exception: Option[Exception] = None
+  private[this] var backendPort: Int = 0
 
-  override def getPartitions = parent.partitions
+  override def getPartitions = firstParent.partitions
 
-  override def compute(split: Partition, task: TaskContext): Iterator[T] = {
+  override def compute(split: Partition, task: TaskContext): Iterator[Row] = {
 
     val sessionId: Int = scala.util.Random.nextInt(10000)
     val logger = new Logger("Worker", sessionId)
     val lock: AnyRef = new Object()
 
-    val workerContext = new WorkerContext[T](
-      parent,
-      split,
-      task,
+    val workerContext = new WorkerContext(
+      firstParent.iterator(split, task).toArray,
       lock,
       closure,
       columns,
@@ -42,33 +39,39 @@ class WorkerRDD[T: ClassTag](
     val contextId = JVMObjectTracker.put(workerContext)
     logger.log("is tracking worker context under " + contextId)
 
+    logger.log("initializing backend")
+    val backend: Backend = new Backend()
+
+    /*
+     * initialize backend as worker and service, since exceptions and
+     * terminating the r session should not shutdown the process
+     */
+    backend.setType(
+      true,   /* isService */
+      false,  /* isRemote */
+      true    /* isWorker */
+    )
+
+    backend.setHostContext(
+      contextId
+    )
+
+    backend.init(
+      port,
+      sessionId
+    )
+
+    backendPort = backend.getPort()
+
     new Thread("starting backend thread") {
       override def run(): Unit = {
         try {
           logger.log("starting backend")
-          val backend: Backend = new Backend()
 
-          /*
-           * initialize backend as worker and service, since exceptions and
-           * closing terminating the r session should not shutdown the process
-           */
-          backend.setType(
-            true,   /* isService */
-            false,  /* isRemote */
-            true    /* isWorker */
-          )
-
-          backend.setHostContext(
-            contextId
-          )
-
-          backend.init(
-            port,
-            sessionId
-          )
+          backend.run()
         } catch {
           case e: Exception =>
-            logger.logError("failed to start backend: ", e)
+            logger.logError("failed while running backend: ", e)
             exception = Some(e)
             lock.synchronized {
               lock.notify
@@ -83,7 +86,7 @@ class WorkerRDD[T: ClassTag](
           logger.log("is starting rscript")
 
           val rscript = new Rscript(logger)
-          rscript.init(sessionId, config)
+          rscript.init(sessionId, backendPort, config)
           lock.synchronized {
             lock.notify
           }
@@ -108,6 +111,7 @@ class WorkerRDD[T: ClassTag](
       throw exception.get
     }
 
+    logger.log("is returning RDD iterator with " + workerContext.getResultArray().length + " rows")
     return workerContext.getResultArray().iterator
   }
 }